+++ app/upgrades.go
@@ -59,7 +59,6 @@ func (app *UmeeApp) registerUpgrade5_1(upgradeInfo upgradetypes.Plan) {
 	planName := "v5.1"
 	app.UpgradeKeeper.SetUpgradeHandler(planName,
 		func(ctx sdk.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
-
 			// GravityBridge is deleted after v5.1
 			// if err := app.GravityKeeper.MigrateFundsToDrainAccount(
 			// 	ctx,