+++ x/ugov/keeper/keys.go
@@ -2,5 +2,6 @@ package keeper
 
 // store key prefixes
 var (
-	keyMinGasPrice = []byte{0x01}
+	keyMinGasPrice    = []byte{0x01}
+	keyEmergencyGroup = []byte{0x02}
 )